+++ campus-booking/src/main/java/com/campus/booking/repository/BookingRepository.java
@@ -3,16 +3,17 @@ package com.campus.booking.repository;
 import com.campus.booking.domain.Booking;
 import com.campus.booking.domain.Room;
 import com.campus.booking.domain.Student;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
+
 import java.util.Optional;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 
 
 @Repository
-public interface BookingRepository extends JpaRepository<Booking, Integer> {
+public interface BookingRepository extends JpaRepository<Booking, Long> {
 
     @Query("""
         SELECT b
