+++ campus-booking/src/main/java/com/campus/booking/repository/RoomRepository.java
@@ -3,10 +3,11 @@ package com.campus.booking.repository;
 import com.campus.booking.domain.Room;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+
 import java.util.Optional;
 
 @Repository
-public interface RoomRepository extends JpaRepository<Room, Integer> {
+public interface RoomRepository extends JpaRepository<Room, Long> {
     Optional<Room> findByRoomCode(String roomCode);
 }
