+++ campus-booking/src/main/java/com/campus/booking/service/impl/RoomServiceImpl.java
@@ -3,12 +3,11 @@ package com.campus.booking.service.impl;
 import com.campus.booking.domain.Room;
 import com.campus.booking.repository.RoomRepository;
 import com.campus.booking.service.RoomService;
-import org.springframework.transaction.annotation.Transactional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 import java.util.Optional;
 
 @Service
@@ -28,7 +27,7 @@ public class RoomServiceImpl implements RoomService {
 
     @Override
     @Transactional(readOnly = true)
-    public Optional<Room> getRoomById(Integer id) {
+    public Optional<Room> getRoomById(Long id) {
         return roomRepository.findById(id);
     }
